// src/utils/enums/TooltipEnum.ts
export enum ButtonType {
    TRIPLE_DOTS = "triple_dots",
    QUESTION_MARK = "question_mark",
    QUESTION_MARK_WHITE = "question_mark_white",
    WATER_DROP = "water_drop"
}

export enum ContainerType {
    DEFAULT = "default",
    CONTAINER = "container"
}

export enum BubbleType {
    DEFAULT = "default",
    NONE = "none"
}

export enum PositionType {
    RIGHT = "right",
    LEFT = "left",
    TOP = "top",
    BOTTOM = "bottom"
}